requirements.txt
Python=3.9.0 (3.8+)
Django=4.2.10
djangorestframework=3.14.0
Pillow=10.2.0
